backend/cloud-api/requirements.txt
yandexcloud
grpcio-tools
psycopg2-binary